refactor: organized authorization classes
