--- backend/Authorization/CommentCreateUpdateRequirement.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNetCore.Authorization;
-
-namespace Backend.Authorization
-{
-    public class CommentCreateUpdateRequirement : IAuthorizationRequirement
-    {
-    }
-}
