+++ backend/Authorization/CommentPolicies/CommentCreateRequirement.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Authorization;
+
+namespace Backend.Authorization.CommentPolicies
+{
+    public class CommentCreateRequirement : IAuthorizationRequirement
+    {
+    }
+}
