+++ backend/Authorization/CommentPolicies/CommentDeleteRequirement.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 
-namespace Backend.Authorization
+namespace Backend.Authorization.CommentPolicies
 {
     public class CommentDeleteRequirement : IAuthorizationRequirement
     {
