--- backend/Authorization/ContentCreateUpdateRequirement.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNetCore.Authorization;
-
-namespace Backend.Authorization
-{
-    public class ContentCreateUpdateRequirement : IAuthorizationRequirement
-    {
-    }
-}
