+++ backend/Authorization/ContentPolicies/ContentCreateRequirement.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Authorization;
+
+namespace Backend.Authorization.ContentPolicies
+{
+    public class ContentCreateRequirement : IAuthorizationRequirement
+    {
+    }
+}
