+++ backend/Authorization/ContentPolicies/ContentDeleteRequirement.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 
-namespace Backend.Authorization
+namespace Backend.Authorization.ContentPolicies
 {
     public class ContentDeleteRequirement : IAuthorizationRequirement
     {
