+++ backend/Authorization/UserPolicies/UserDeleteRequirement.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 
-namespace Backend.Authorization
+namespace Backend.Authorization.UserPolicies
 {
     public class UserDeleteRequirement : IAuthorizationRequirement
     {
