+++ backend/Authorization/UserPolicies/UserRequirement.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Authorization;
+
+namespace Backend.Authorization.UserPolicies
+{
+    public class UserRequirement : IAuthorizationRequirement
+    {
+    }
+}
