+++ backend/Controllers/UserController.cs
@@ -43,14 +43,12 @@ namespace InstaConnect.Controllers
             return await _userService.CreateUsersAsync(newUsers);
         }
 
-        [Authorize(Policy = "UserPolicy")]
         [HttpPut("User")]
         public async Task<ActionResult<UserModel>> PutUser([FromBody] UserModel? newUser)
         {
             return await _userService.UpdateUserAsync(newUser);
         }
 
-        [Authorize(Policy = "UserPolicy")]
         [HttpPut("Users")]
         public async Task<ActionResult<List<UserModel>>> PutUsers([FromBody] List<UserModel>? newUsers)
         {
